requirements.txt
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
aws-xray-sdk==0.95
awscli==1.16.309
backcall==0.1.0
bcrypt==3.1.7
boto==2.49.0
boto3==1.7.84
botocore==1.13.45
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.3
cookies==2.2.1
cryptography==2.8
decorator==4.4.0
defusedxml==0.6.0
django-allauth==0.39.1
django-cors-headers==3.2.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-paypal==1.0.0
django-s3-storage==0.13.0
django-storages==1.8
docker==4.1.0
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
geolocation-python==0.2.2
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
moto==1.3.6
oauthlib==3.0.1
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyaml==19.12.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.4.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pypiwin32==223
python-dateutil==2.8.1
python-decouple==3.1
python-jose==2.0.2
python3-openid==3.1.0
pytz==2018.5
pywin32==227
PyYAML==5.2
requests==2.21.0
requests-oauthlib==1.2.0
responses==0.10.9
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.2.4
stripe==2.42.0
timeago==1.0.10
traitlets==4.3.2
typed-ast==1.4.1
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
